:sparkles: 添加active_updater 相关 trait
